fix: Improve watcher script
Add a wait condition in case the wayland socket is not ready yet. Using the DELETE_SELF event on the socket itself lead to race conditions where inotifywait seemed to listen, but probably to an outdated inode. The sleep delay should ensure that the browser is restarted after mir-kiosk is up again.

Signed-off-by: Tobias Grasse <tg@glancr.de>